backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for realtime app"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-socketio = "^0.1.1"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.0"
"python-socketio" = {extras = ["asyncio"], version = "^5.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.1"
httpx = "^0.23.0"
async-asgi-testclient = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 